Allow browsing current correspondence as well as whole volume
- Change header layout
- Replace "dumb" components with partials
- Move Solr function to mixin
- Load basic metadata for all letters on first load

Updated Ember CLI to 1.13.13 that fixed a bug related to Babel
polyfills which in turn was necessary for the new tests to run.

Resolves: ADWD-2000, ADWD-2001
