+++ app/components/header-container.js
@@ -1,4 +1,5 @@
 import Ember from 'ember';
 
 export default Ember.Component.extend({
+  tagName: ''
 });
