+++ app/components/letter-header.js
@@ -3,7 +3,11 @@ import Ember from 'ember';
 export default Ember.Component.extend({
   showCitation: false,
   showMetadata: false,
+  tagName: '',
   actions: {
+    changeLetter(id) {
+      this.sendAction('changeLetter', id);
+    },
     toggleCitation() {
       Ember.$('.citation').slideToggle();
       this.set('showCitation', ! this.get('showCitation'));
