--- app/components/letter-metadata.js
@@ -1,4 +1,0 @@
-import Ember from 'ember';
-
-export default Ember.Component.extend({
-});
