+++ app/styles/blocks/header.scss
@@ -4,6 +4,10 @@
   float: left;
   position: relative;
   width: 100%;
+
+  p:nth-child(n+2) {
+    margin-top: ($g / 4);
+  }
 }
 
 .header_logo {
@@ -34,27 +38,21 @@
 
 .header_main {
   height: (3 * $g);
-  padding: ($g / 2) 12em ($g / 2) (5 * $g); // Make room on right for header_aside
+  padding: ($g / 4) ($g / 2) ($g / 4) (5 * $g); // Make room on left for logo
 }
 
-.header_aside {
-  padding: ($g / 2);
-  position: absolute;
-  right: 0;
-  top: 0;
+.header_pagination {
+  float: right;
+  text-align: right;
 }
 
 .header_title {
-  @include typo-scale(1);
-  margin: 0;
+  float: left;
 }
 
 .header_correspondents {
-  display: inline-block;
-}
-
-.header_dates {
-  display: inline-block;
+  @include typo-scale(1);
+  margin-top: floor($g / 8);
 }
 
 .header_date {
@@ -63,10 +61,6 @@
 
 .header_step {
   @include button('shady');
-
-  @include hover {
-    background: $shade;
-  }
 
   &.-prev {
     @include icon('chevron-left', false, 1);
