+++ app/styles/mixins/button.scss
@@ -5,32 +5,26 @@
   cursor: pointer;
   display: inline-block;
   font: inherit;
-  min-width: (1.5 * $g);
-  outline: 0;
+  min-width: $g;
   padding: 0 ($g / 4);
   text-align: center;
 
-  @include hover {
-    transition: none;
-  }
-
   @if $style == 'shady' {
+    background: $shade;
     box-shadow: none;
     color: $link-color;
-
-    @include hover {
-      background: $shade;
-      box-shadow: 0 (-$border-width) 0 $shade-medium inset;
-      color: $link-hover-color;
-    }
   } @else {
     background: $link-color;
     box-shadow: 0 (-$border-width) 0 $shade-medium inset;
     color: auto-color($link-color);
     text-shadow: 0 -1px 0 $shade-medium;
+  }
 
-    @include hover {
-      background: $link-hover-color;
-    }
+  @include hover {
+    background: $link-hover-color;
+    box-shadow: 0 (-$border-width) 0 $shade-medium inset;
+    color: auto-color($link-hover-color);
+    text-shadow: 0 -1px 0 $shade-medium;
+    transition: none;
   }
 }
