+++ app/templates/letter.hbs
@@ -1,7 +1,7 @@
 {{#if content.brief_nummer}}
 
   {{#header-container}}
-    {{letter-header content=content}}
+    {{letter-header content=content changeLetter='changeLetter'}}
   {{/header-container}}
 
   <main class="content" id="content">
