+++ app/templates/partials/letter-citation.hbs
@@ -1,14 +1,17 @@
 <aside class="citation">
   <h2 class="citation_heading">{{t 'citation'}}</h2>
 
+  {{!TODO: "This is not the complete edition"}}
+  {{!TODO: "The complete edition can be found here:"}}
+
   <span class="citation_overline">{{t 'historicCriticalEdition'}}</span>
   <h2 class="citation_heading">
     <a href="{{content.brief_url}}" target="_blank">Leibniz, Gottfried Wilhelm: Sämtliche Schriften und Briefe</a>
   </h2>
   {{!TODO: Translate}}
   <p>Hrsg. v. Berlin-Brandenburgischen Akademie der Wissenschaften / Akademie der Wissenschaften zu Göttingen</p>
-  <h3 class="citation_sub-heading">{{t 'permaLink'}}</h3>
-  <input class="citation_input" type="url" value="{{content.brief_url}}" readonly>
+
+  {{!TODO: Citation example: <example>}}
 
   {{yield}}
 </aside>
