+++ bower.json
@@ -10,7 +10,7 @@
     "ember-qunit-notifications": "0.1.0",
     "ember-resolver": "~0.1.20",
     "jquery": "^1.11.3",
-    "loader.js": "ember-cli/loader.js#3.3.0",
+    "loader.js": "ember-cli/loader.js#3.4.0",
     "qunit": "~1.20.0",
     "reset-css": "~2.0.2011012602",
     "MathJax": "components/MathJax#~2.5.3",
