+++ ember-cli-build.js
@@ -4,7 +4,10 @@ var Funnel = require('broccoli-funnel');
 
 module.exports = function(defaults) {
   var app = new EmberApp(defaults, {
-    // Add options here
+    babel: {
+      optional: ['es6.spec.symbols'],
+      includePolyfill: true
+    }
   });
 
   // Use `app.import` to add additional libraries to the generated
