+++ package.json
@@ -21,7 +21,7 @@
   "devDependencies": {
     "broccoli-asset-rev": "^2.3.0",
     "broccoli-funnel": "^1.0.0",
-    "ember-cli": "1.13.12",
+    "ember-cli": "1.13.13",
     "ember-cli-app-version": "^1.0.0",
     "ember-cli-autoprefixer": "0.5.0",
     "ember-cli-babel": "^5.1.6",
@@ -35,7 +35,7 @@
     "ember-cli-qunit": "^1.0.4",
     "ember-cli-release": "0.2.8",
     "ember-cli-sass": "^5.1.0",
-    "ember-cli-sri": "^1.1.0",
+    "ember-cli-sri": "^1.2.0",
     "ember-cli-uglify": "^1.2.0",
     "ember-data": "2.1.0",
     "ember-disable-proxy-controllers": "^1.0.1",
