+++ tests/acceptance/toggle-variants-test.js
@@ -8,7 +8,7 @@ test('toggle variants', function(assert) {
   assert.expect(2);
 
   // Make sure this letter actually has variants
-  visit('/letter/l3717');
+  visit('/');
 
   var variantsCount = 0
   var highlightsCount = 0
@@ -20,7 +20,7 @@ test('toggle variants', function(assert) {
     highlightsCount = find('.transcript .reference.-highlight').length;
   })
 
-  click('.variants .variants_button:first');
+  click('.variants .variants_button');
   andThen(function() {
     var visibleVariantsCount = find('.variants .variant:visible').length;
     var lessHighlightsCount = find('.transcript .reference.-highlight').length;
