+++ translations/de-de.yaml
@@ -1,10 +1,11 @@
 aboutThisProjekt: 'Über das Projekt'
 addedByEditor: 'Vom Herausgeber hinzugefügt'
+attachment: 'Beilage:'
 attachmentTo: 'Beilage zu'
 browse: 'Blättern'
 citation: 'Zitierfähige Vollfassung'
-citationHide: 'Zitierfähige Vollfassung verbergen'
-citationShow: 'Zitierfähige Vollfassung anzeigen'
+citationHide: 'Zitiervorschlag verbergen'
+citationShow: 'Zitiervorschlag anzeigen'
 contact: 'Kontakt'
 context: 'Kontext'
 databaseError: 'Datenbankfehler'
@@ -33,11 +34,11 @@ languageName:
   rus: 'Russisch'
   swe: 'Schwedisch'
 leibnizLetters: 'Briefportal Leibniz'
-letterNext: 'Folgender Brief dieser Korrespondenz'
+letterNext: 'Folgender Brief'
 letterNotExisting: 'Brief <i>{id}</i> existiert nicht.'
 letterNotExistingPrev: 'Kein vorhergehender Brief vorhanden'
-letterNotExistingNext: 'Kein Nachfolgender Brief vorhanden'
-letterPrev: 'Vorhergehender Brief dieser Korrespondenz'
+letterNotExistingNext: 'Kein nachfolgender Brief vorhanden'
+letterPrev: 'Vorhergehender Brief'
 loading: 'Brief wird geladen &hellip;'
 location: 'Ort'
 locations: 'Orte'
@@ -55,6 +56,10 @@ pageNotFound: 'Verzeihung, diese Seite ist nicht vorhanden.'
 pagePrev: 'vorige Seite'
 permaLink: 'Persistente URL:'
 recipient: 'Empfänger'
+respondsIsFirst: 'Erster Brief dieser Korrespondenz'
+respondsTo: 'Antwort auf'
+respondedBy: 'beantwortet durch'
+respondsIsLast: 'letzter Brief dieser Korrespondenz'
 restore: 'Wieder herstellen'
 search: 'Suchen'
 searchLetters: 'Briefe suchen'
